Implementing add method.

+++ src/main/java/com/springboot/todoWebApp/dao/ToDoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ToDoRepository extends JpaRepository<ToDo, Long> {
+public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
 }
